Add files via upload
Dog Attacks  
Tasks  
1.  How many attacks occurred per year?
Create a histogram to represent the attacks per year
2.  How many attacks occured per month per year?
Create one line in which for each year theree is a different line.
3.  What are the three states in which the most attacks occurred between 2020 and 2022.
Create one histogram to represent the data
4.  For each of the following age groups, how many people were attacked. 
Age groups
minor: up to 17 years old
adult: 18 to 65 years old 
older adult: 65 and up
Create a histogram showing the results
5.  How many attacks were performed by each breed?
Show the data using a histogram
Notes  
Think on how you can create functions or classes to create a code that is reusable, 
modular, maintainable, and generic to work with any csv file related to dog attack data.
Presentation  
Each student is required to make a demo showcasing the solution and explaining the 
reasons behind the design decisions regarding the code. That is, why you decided to 
structure the code the way you did.
The data in the CSV file is not fully cleaned. You will see inconsistancies. There is no right or 
wrong way of handling them. Part of the assignment is for you to handle them and during 
the presentation explain how you handled it and why you chose that method.
Be prepared for other students and the instructor to ask questions regarding your code